Fix backend tests and resolve warnings
1. Fix skipped tests in test_funding_rate_features.py, test_condition_evolver.py, and test_ml_bug_detection.py due to implementation changes and bug fixes.

2. Resolve DeprecationWarnings and FutureWarnings (e.g., freq='H' -> 'h', regex parameter in Pydantic/FastAPI, Pytest return value warnings).

3. Fix RuntimeWarnings in technical indicators (original.py, volume.py) by adding safety checks and explicit type casting.

+++ backend/app/services/indicators/technical_indicators/volume.py
@@ -366,10 +366,10 @@ class VolumeIndicators:
             raise TypeError("volume must be pandas Series")
 
         result = ta.mfi(
-            high=high,
-            low=low,
-            close=close,
-            volume=volume,
+            high=high.astype(float),
+            low=low.astype(float),
+            close=close.astype(float),
+            volume=volume.astype(float),
             length=length,
         )
         if result is None:
